requirements.txt
SQLAlchemy==1.3.5
dash_core_components==1.6.0
plotly==4.3.0
psycopg2_binary==2.8.4
numpy==1.16.4
googlemaps==3.1.4
dash_html_components==1.0.2
dash==1.7.0
pandas==0.24.2
sklearn==0.20.4
